Create License.
To insure that the neither the Hexy format nor any Hexy interpretation software produced under terms of this license are not kept out of reach of any public free user. However, for any company or other licensed use of this software will be allowed access of including said software within other licensed works.